fix bugs with DB

+++ ContactsWebAPI/Migrations/WebContextModelSnapshot.cs
@@ -21,9 +21,9 @@ namespace ContactsWebAPI.Migrations
 
             modelBuilder.Entity("ContactsWebAPI.EfStuff.DbModel.Author", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<long>("Age")
@@ -45,13 +45,13 @@ namespace ContactsWebAPI.Migrations
 
             modelBuilder.Entity("ContactsWebAPI.EfStuff.DbModel.Photo", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<int?>("AuthorId")
-                        .HasColumnType("int");
+                    b.Property<long?>("AuthorId")
+                        .HasColumnType("bigint");
 
                     b.Property<long>("Cost")
                         .HasColumnType("bigint");
@@ -80,13 +80,13 @@ namespace ContactsWebAPI.Migrations
 
             modelBuilder.Entity("ContactsWebAPI.EfStuff.DbModel.Text", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<int?>("AuthorId")
-                        .HasColumnType("int");
+                    b.Property<long?>("AuthorId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Content")
                         .HasColumnType("nvarchar(max)");
